New Makefile for the examples.
git-svn-id: https://svn-eos.cs.uni-magdeburg.de/repos/Projects/AVR/halib/trunk@444 ac1e7a95-1e17-0410-ad1e-f2b06e490645

+++ examples/applications/uadc.cpp
@@ -70,7 +70,7 @@ int main()
 	while(true)
 	{
 		for (int i = 0; i<6; i++)
-			uart << getValue(i) << "\t";
+			uart << (int32_t)getValue(i) << "\t";
 
 #if 0			
 		uart << getValue(0) << "\t";
